chore(compiler): refactoring class field initialization
